batch-process-docs/pyproject.toml
[project]
name = "batch-process-docs-runtime"
version = "0.1.0"
description = "Batch document processing workflow using Google Gemini to extract information from documents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exospherehost>=0.0.2",
    "python-dotenv>=1.1.1",
    "google-genai>=0.3.0",
    "requests>=2.28.0",
    "pydantic>=1.10.0",
    "aiohttp>=3.8.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "jsonschema>=4.0.0",
    "motor>=3.7.1",
    "pypdf2>=3.0.1",
    "google-genai>=1.33.0",
]
